+++ src/clarity/__init__.py
@@ -6,7 +6,7 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from sphinx.application import Sphinx
 
-__version__ = "1.0.0"
+__version__ = "1.0.1"
 
 
 def get_html_theme_path() -> str: